Refactor MMENGINE UTILS 2
