--- siatune/mm/core/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) SI-Analytics. All rights reserved.
-from .utils import IS_DEPRECATED_MMCV
-
-__all__ = [
-    'IS_DEPRECATED_MMCV',
-]
